FP in AnnotationNewLineRule with comment
### What's done:
   Fixed according our code-style
