+++ diktat-rules/src/main/kotlin/org/cqfn/diktat/ruleset/generation/Generation.kt
@@ -25,26 +25,8 @@ private val autoGenerationComment =
             | This document contains all enum properties from Warnings.kt as Strings.
         """.trimMargin()
 
-private val autoTable = """
-\begin{center}
-\scriptsize
-\begin{longtable}{ |l|p{0.8cm}|p{0.8cm}| p{3cm} | }
-\hline
-\multicolumn{4}{|c|}{Available Rules} \\ 
-\hline
-\textbf{diKTat rule} & \textbf{code style} & \textbf{autofix} &  \textbf{config} \\
-\hline
-""".trimIndent()
-
-private val autoEnd = """
-\hline
-\end{longtable}
-\end{center}
-""".trimIndent()
-
 fun main() {
     generateWarningNames()
-    generateAvailableRules()
     validateYear()
 }
 
@@ -72,37 +54,6 @@ private fun generateWarningNames() {
         .build()
 
     kotlinFile.writeTo(File("diktat-rules/src/main/kotlin"))  // fixme: need to add it to pom
-}
-
-@Suppress("MagicNumber")
-private fun generateAvailableRules() {
-    val ruleMap = File("info/rules-mapping.md").readLines().drop(2)
-            .map { it.drop(1).dropLast(1).split("|") }
-            .map { RuleDescription(it[0].replace("\\s+".toRegex(), ""), it[1], it[2], null) }
-            .map { it.ruleName to it }.toMap()
-    File("info/available-rules.md").readLines()
-            .drop(2)
-            .map { it.drop(1).dropLast(1).split("|") }
-            .map { it[2].replace("\\s+".toRegex(), "") to it[5] }
-            .forEach { ruleMap[it.first]?.config = it.second }
-    val newText = File("wp/sections/appendix.tex").readLines().toMutableList()
-    newText.removeAll(newText.subList(newText.indexOf("\\subsection{available-rules}") + 1,newText.indexOf("\\subsection{guide}")))
-    var index = newText.indexOf("\\subsection{available-rules}") + 1
-    autoTable.split("\n").forEach { newText.add(index++, it) }
-    ruleMap.map { it.value }
-            .map { "${it.correctRuleName} & ${it.correctCodeStyle} & ${it.autoFix} & ${it.correctConfig}\\\\" }
-            .forEach { newText.add(index++, it) }
-    autoEnd.split("\n").forEach { newText.add(index++, it) }
-    File("wp/sections/appendix.tex").writeText(newText.joinToString(separator = "\n"))
-}
-
-@Suppress("UnsafeCallOnNullableType")
-data class RuleDescription(val ruleName: String, val codeStyle: String, val autoFix: String, var config: String?) {
-    val correctCodeStyle = codeStyle.substring(codeStyle.indexOf("[") + 1, codeStyle.indexOf("]"))
-    val correctRuleName = ruleName.replace("_", "\\underline{ }")
-    val correctConfig: String by lazy {
-        config!!.replace("<br>", " ")
-    }
 }
 
 private fun validateYear() {
