+++ info/available-rules.md
@@ -23,6 +23,7 @@
 |    1  |       1.2  |  FILE_NAME_MATCH_CLASS                    |  Check: warns<br>Fix: no                                                                                                                                                       |   no  |  no  |  Probably it can be autofixed, but it will be aggressive fix|
 |    1  |       1.2  |  EXCEPTION_SUFFIX                         |  Check: warns if class that extends any Exception class does not have Exception suffix<br>Fix: Adding suffix "Exception" to a class name                                       |  yes  |  no  | Need to add tests for this|
 |    2  |       2.1  |  MISSING_KDOC_TOP_LEVEL                   |  Check: warns on file level internal or public class or function has missing KDoc<br>Fix: no                                                                                   |   no  |  no  | Support extension for setters/gettersSupport extension for method "override"|
+|    2  |      2.1.1 |  KDOC_EXTRA_PROPERTY                      |  Check: warn if there is property in KDoc which is not present in the class                                                                                                    |   no  | no   |  - |
 |    2  |       2.1  |  MISSING_KDOC_CLASS_ELEMENTS              |  Check: warns if accessible internal elements (protected, public, internal) in a class are not documented<br>Fix: no                                                           |   no  |  no  | May be need to add exception cases for setters and getters. There is no sense in adding KDoc to them.|
 |    2  |       2.1  |  MISSING_KDOC_ON_FUNCTION                 |  Check: warns if accessible function doesn't have KDoc<br>Fix: adds KDoc template if it is not empty                                                                           |  yes  |  no  |  |
 |    2  |       2.2  |  KDOC_WITHOUT_PARAM_TAG                   |  Check: warns if accessible method has parameters and they are not documented in KDoc<br>Fix: If accessible method has no KDoc, KDoc template is added                         |  yes  |  no  | Should also make separate fix, even if there is already some Kdoc in place|
@@ -101,9 +102,9 @@
 |    6  |      6.1.6  |  CLASS_SHOULD_NOT_BE_ABSTRACT            |  Checks: if abstract class has any abstract method. If not, warns that class should not be abstract<br>Fix: deletes abstract modifier                                          |  yes  |  no  |  - |
 |    6  |      6.1.7  |  NO_CORRESPONDING_PROPERTY               |  Checks: if in case of using "backing property" scheme, the name of real and back property are the same                                                               |  no  |  no  |  - |
 |    6  |      6.1.9  |  WRONG_NAME_OF_VARIABLE_INSIDE_ACCESSOR  |  Check: used the name of a variable in the custom getter or setter                                                                                                             |   no  |  no |
-|    6  |     6.1.10  |  TRIVIAL_ACCESSORS_ARE_NOT_RECOMMENDED   |  Check: if there are any trivial getters or setters <br> Fix: Delete trivial getter or setter                                                                                  |  yes  |  no  |  - |
-|    6  |     6.1.8   |  CUSTOM_GETTERS_SETTERS                  |  Check: Inspection that checks that no custom getters and setters are used for properties                                                                                      |  no  |  no  |  - |
-|    6  |     6.1.11  |  COMPACT_OBJECT_INITIALIZATION           |  Checks if class instantiation can be wrapped in `apply` for better readability  |    |  no  |  |
+|    6  |      6.1.10 |  TRIVIAL_ACCESSORS_ARE_NOT_RECOMMENDED   |  Check: if there are any trivial getters or setters <br> Fix: Delete trivial getter or setter                                                                                  |  yes  |  no  |  - |
+|    6  |      6.1.8  |  CUSTOM_GETTERS_SETTERS                  |  Check: Inspection that checks that no custom getters and setters are used for properties                                                                                      |  no  |  no  |  - |
+|    6  |      6.1.11 |  COMPACT_OBJECT_INITIALIZATION           |  Checks if class instantiation can be wrapped in `apply` for better readability  |    |  no  |  |
 |    6  |      6.2.2  |  EXTENSION_FUNCTION_SAME_SIGNATURE       |  Checks if extension function has the same signature as another extension function and their classes are related                                                               |  no  |  no  |  + |
-|    6  |     6.4.1   | AVOID_USING_UTILITY_CLASS                | Checks if there is class/object that can be replace with extension function                                                                                                  | no  | - | - |
-|    6  |     6.4.2   | OBJECT_IS_PREFERRED                      | Checks: if class is stateless it is preferred to use `object`                                                                                                                | yes | - | + |
+|    6  |      6.4.1  |  AVOID_USING_UTILITY_CLASS               | Checks if there is class/object that can be replace with extension function                                                                                                  | no  | no | - |
+|    6  |      6.4.2  |  OBJECT_IS_PREFERRED                     | Checks: if class is stateless it is preferred to use `object`                                                                                                                | yes | no | + |
